// src/models/register.ts
export interface Register {
  firstName: string;
  lastName: string;
  email: string;
  contactNumber: string;
  password: string;
  confirmPassword: string;
  userType: string;
  serviceType: string;
  latitude: string;
  longitude: string;
  address: string;
  stageName: string;
  city: string;
  state: string;
  country: string;
}